Add Makefile-based workflow with CLI argument-based tasks
This adds a basic workflow for doing object recognition on MNIST and CIFAR10
using PyTorch on CPU. Each task is a program with CLI arguments and tasks are
connected using a Makefile. Running the workflow results in a directory for each
task that was run. Ultimately, Makefiles probably aren't the right solution, but
I wanted to try something simple and classic first.